+++ runners/handlers/bytes-iterable.ts
@@ -48,6 +48,12 @@ export class BytesIterableInputHandler
         await bw.write(byteArray);
       }
       await bw.flush();
+    } catch (e) {
+      if (e instanceof Deno.errors.BrokenPipe) {
+        // Ignore.
+      } else {
+        throw e;
+      }
     } finally {
       stdin.close();
     }
@@ -70,7 +76,7 @@ export class BytesIterableOutputHandler
     stdout: MultiCloseReader,
     stderr: MultiCloseReader,
     process: MultiCloseProcess,
-    input: Promise<void>,
+    input: { stdin: MultiCloseWriter; handlerResult: Promise<void> },
   ): AsyncIterable<Uint8Array> {
     return this.process(stdout, stderr, process, input);
   }